+++ CHANGELOG.md
@@ -1,2 +1,1 @@
-- Added support for Fabric and Forge 1.20.x.
-- Internal changes. This should not affect anything in the mod in any way. Please report any bugs or discrepancies at the issue tracker.
+- Fixed crash on Forge.